fix weird react state update error message
Message was: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.

This was when I tried calling setUser in the sign in handler. Just ended up reverting the offending code.